Use a dedicated file watcher
This change introduces a single, pure filesystem watcher that dispatches change
events. The live-reload and ingest systems instead subscribe to events from the
fs channel and translates relevant messages to "app events" which the live
reload handler can use to control the client.

This change also paves the way for specifically handling changes to the db
schema.