Refactor: Streamline app and dev concepts
This commit removes the reliance on integrant to build the structure of
dependencies - this only works in development and causes inconsistencies in the
export.

This commit introduces powerpack.app as a way to instrument and start the main
powerpack app, and then uses integrant only to wire up the development
environment. This should both make the code easier to read (no more bespoke
combinations of dependencies, and all functions receive the same basic two
constructs), as well as make the export work better with less code.